Add SelectMultiple to ISelectableToken; Organize ISelectableToken interface and classes
