+++ JUST.net/DataTransformer.cs
@@ -5,12 +5,12 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Text;
 using System.Linq;
+using JUST.net.Selectables;
 
 namespace JUST
 {
     public class DataTransformer : DataTransformer<JsonPathSelectable>
     {
-
     }
 
     public class DataTransformer<T> where T: ISelectableToken
